+++ app/models/timeline/renderers/svg/rendererable_time_rage_event.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module Timeline
+  class RendererableTimeRangeEvent < RenderableEvent
+  end
+end
