+++ spec/support/test_factories/db/timeline_factory.rb
@@ -8,7 +8,6 @@ class TestFactories
         name: "Timeline-#{seq}",
         slug: "timeline-#{seq}",
         description: 'Some Description',
-        color: "color-#{}",
       }.merge(overrides)
     end
   end
